+++ frontend/public/js/fileDelete.js
@@ -1,12 +1,12 @@
-import { $deleteFile, $fileSelector } from "./components.js";
+import { deleteFile, fileSelector } from "./components.js";
 
-if (!$deleteFile.length) {
+if (!$(deleteFile).length) {
   console.error("Delete file button not found");
 }
 
 const handleDeleteFile = async () => {
   if (confirm("Are you sure? This cannot be undone.")) {
-    const fileID = $fileSelector.val();
+    const fileID = $(fileSelector).val();
     if (!fileID) {
       console.error("Invalid file ID");
       return;
@@ -14,17 +14,10 @@ const handleDeleteFile = async () => {
     const response = await fetch(`/delete/${fileID}`, { method: "DELETE" });
     console.log(response);
     if (response.ok) {
-      if ($(`#${fileID}`).length) 
-        $(`#${fileID}`).remove();
+      if ($(`#${fileID}`).length) $(`#${fileID}`).remove();
     }
     $(document).trigger("complete");
   }
 };
 
-$deleteFile.on("click", handleDeleteFile);
-
-export const deleteFileButton = () =>
-  $("<button>", {
-    text: "Delete",
-    class: "dropdown-item",
-  }).on("click", handleDeleteFile);
+$(deleteFile).on("click", handleDeleteFile);