+++ frontend/public/js/fileDump.js
@@ -1,12 +1,12 @@
-import { $dumpFile, $fileSelector } from "./components.js";
+import { dumpFile, fileSelector } from "./components.js";
 
-if (!$dumpFile.length) {
+if (!$(dumpFile).length) {
   console.error("Dump file button not found");
 }
 
 const handleDumpFile = async () => {
   if (confirm("Do you want to move this file to trash?")) {
-    const fileID = $fileSelector.val();
+    const fileID = $(fileSelector).val();
     if (!fileID) {
       console.error("Invalid file ID");
       return;
@@ -20,10 +20,4 @@ const handleDumpFile = async () => {
   }
 };
 
-$dumpFile.on("click", handleDumpFile);
-
-export const dumpFileButton = () =>
-  $("<button>", {
-    text: "Delete",
-    class: "dropdown-item",
-  }).on("click", handleDumpFile);
+$(dumpFile).on("click", handleDumpFile);