+++ frontend/public/js/filePanel.js
@@ -1,15 +1,19 @@
-import { $contextMenu, $filePanel, $fileSelector, $renameFile } from "./components.js";
-import { appendItem, clearMenu, showMenuAtPos } from "./contextMenu.js";
+import {
+  contextMenu,
+  filePanel,
+  fileSelector
+} from "./components.js";
+import { showMenuAtPos } from "./contextMenu.js";
 
-if (!$contextMenu.length) {
+if (!$(contextMenu).length) {
   console.error("Context menu not found");
 }
 
-if (!$filePanel.length) {
+if (!$(filePanel).length) {
   console.error("File panel not found");
 }
 
-if (!$fileSelector.length) {
+if (!$(fileSelector).length) {
   console.error("File selector not found");
 }
 
@@ -18,21 +22,11 @@ const handleFileSelect = (e) => {
   const target = e.target.closest(".id-selectable");
   if (!target) return;
   const fileID = target.id;
-  $contextMenu.find("a.read").attr("href", `/read/${fileID}`);
-  $fileSelector.val(fileID);
+  $(contextMenu).find("a.read").attr("href", `/read/${fileID}`);
+  $(fileSelector).val(fileID);
   showMenuAtPos({ top: e.clientY, left: e.clientX });
 };
 
-$filePanel
+$(filePanel)
   .on("contextmenu", handleFileSelect)
   .on("click", ".option", handleFileSelect);
-
-$renameFile.on("click", () => {
-  const fileID = $fileSelector.val();
-  if (!fileID) {
-    console.error("Invalid file ID");
-    return;
-  }
-  $contextMenu.trigger("complete");
-  $(`#${fileID} .title`).trigger("focus");
-});