+++ frontend/public/js/imageDelete.js
@@ -1,41 +1,18 @@
-import { $fileSelector, $imageSelector } from "./components.js";
+import { clearImages, fileSelector, imageSelector } from "./components.js";
 import { saveFile } from "./fileSave.js";
 
-export const deleteLink = (anchor) => {
-  if (!anchor || !anchor.href) {
-    console.error("Invalid anchor element");
-    return;
-  }
-  $(anchor).replaceWith($(anchor).html());
-};
-
-export const deleteLinkByImage = (imageID) => {
-  console.log(imageID);
-  const basename = imageID.split("/").pop();
-  $(`a[href$="${basename}"]`).each((index, anchor) => {
-    console.log(anchor);
-    $(anchor).replaceWith($(anchor).html());
-  });
-};
-
-if (!$fileSelector.length) {
+if (!$(fileSelector).length) {
   console.error("File selector not found");
 }
 
-if (!$imageSelector.length) {
+if (!$(imageSelector).length) {
   console.error("Image selector not found");
 }
 
-export const deleteImageButton = () =>
-  $("<button>", {
-    text: "Delete Image",
-    class: "dropdown-item",
-  }).on("click", handleDeleteImage);
-
 const handleDeleteImage = async () => {
   if (confirm("Are you sure? All links to this image will be removed.")) {
-    const fileID = $fileSelector.val();
-    const imageID = $imageSelector.val();
+    const fileID = $(fileSelector).val();
+    const imageID = $(imageSelector).val()?.toString();
     if (!fileID || !imageID) {
       console.error("Invalid file or image ID");
       return;
@@ -44,10 +21,45 @@ const handleDeleteImage = async () => {
       method: "DELETE",
     });
     if (response.ok) {
-      document.getElementById(imageID).remove();
+      const imageEle = document.getElementById(imageID);
+      if (imageEle) imageEle.remove();
       deleteLinkByImage(imageID);
       saveFile();
     }
     $(document).trigger("complete");
   }
 };
+
+const handleClearImages = async () => {
+  if (
+    confirm(
+      "Are you sure? All images and all links will be removed. This cannot be undone."
+    )
+  ) {
+    const fileID = $(fileSelector).val();
+    if (!fileID) {
+      console.error("Invalid file");
+      return;
+    }
+    const response = await fetch(`/delete/${fileID}/all`, {
+      method: "DELETE",
+    });
+    if (response.ok) {
+      $(".id-selectable[id]").each((index, imageEle) => {
+        const imageID = imageEle.id
+        deleteLinkByImage(imageID);
+        imageEle.remove();
+      });
+      saveFile();
+    }
+    $(document).trigger("complete");
+  }
+};
+
+$(clearImages).on("click", handleClearImages);
+
+export const deleteImageButton = () =>
+  $("<button>", {
+    text: "Delete",
+    class: "dropdown-item delete-image",
+  }).on("click", handleDeleteImage);