+++ frontend/public/js/imageSelect.js
@@ -1,43 +1,52 @@
-import { $editor, $fileSelector } from "./components.js";
+import { editor, fileSelector } from "./components.js";
 import { appendItem, clearMenu, showMenu } from "./contextMenu.js";
-import { deleteLink } from "./imageDelete.js";
+import { deleteLink } from "./components.js";
 import { saveFile } from "./fileSave.js";
 
 const openSelectWindow = async () => {
-  const fileID = $fileSelector.val();
+  const fileID = $(fileSelector).val();
   if (!fileID) {
     console.error("Invalid file ID");
     return;
   }
   const response = await fetch(`/access/${fileID}`);
   const html = await response.text();
-  const $selectWindow = $("<div>", {
+  const $selectionWindow = $("<div>", {
     css: {
       width: "50vw",
     },
     html: html,
   }).on("click", handleSelectImage);
   clearMenu();
-  appendItem($selectWindow);
+  appendItem($selectionWindow);
   showMenu();
 };
 
 const handleSelectImage = (e) => {
   const target = e.target.closest(".id-selectable");
   if (!target) return;
-  const fileID = $fileSelector.val();
+  const fileID = $(fileSelector).val();
   const imageID = target.id;
   if (!fileID || !imageID) {
     console.error("Invalid file or image ID");
     return;
   }
   const selection = window.getSelection();
+  if (!selection) {
+    console.error("No selection found");
+    return;
+  }
+  if (selection.rangeCount === 0) {
+    console.error("No range selected");
+    return;
+  }
   const selectedText = selection.toString();
-  if (selection.rangeCount === 0) console.log("No range selected");
   const range = selection.getRangeAt(0);
-  $editor.find("a.view").each((i, anchor) => {
-    if (range.intersectsNode(anchor)) deleteLink(anchor);
-  });
+  $(editor)
+    .find("a.view")
+    .each((i, anchor) => {
+      if (range.intersectsNode(anchor)) deleteLink(anchor);
+    });
   const $anchor = $("<a>", {
     href: `/access/${fileID}/${imageID}`,
     html: selectedText,
@@ -50,8 +59,7 @@ const handleSelectImage = (e) => {
 
 export const selectImageButton = () => {
   return $("<button>", {
-    id: "select-image",
     text: "Select Image",
-    class: "dropdown-item",
+    class: "dropdown-item select-image",
   }).on("click", openSelectWindow);
 };