+++ frontend/public/js/imageUpload.js
@@ -1,79 +1,97 @@
-import { $fileSelector, $input, $inputForm } from "./components.js";
-import { saveFile } from "./fileSave.js";
+import {
+  contextMenu,
+  fileSelector,
+  imageInput,
+  imagePanel,
+  input,
+  uploadImage,
+} from "./components.js";
 
-if (!$input.length) {
-  console.error("No input element found");
-}
-if (!$inputForm.length) {
-  console.error("No input form found");
+if (!$(contextMenu).length) {
+  console.error("Context menu not found");
 }
 
-if (!$fileSelector.length) {
+if (!$(fileSelector).length) {
   console.error("File selector not found");
 }
 
-const handleUploadImage = async () => {
-  console.log("Uploading image");
-  const selection = window.getSelection();
-  const selectedText = selection.toString();
-  if (selectedText && !/^\s+$/.test(selectedText)) {
-    const range = selection.getRangeAt(0);
-    if (range.startContainer === range.endContainer) {
-      uploadImage().then((imageID) => {
-        const fileID = $fileSelector.val();
-        const $anchor = $("<a>", {
-          href: `/access/${fileID}/${imageID}`,
-          html: selectedText,
-          class: "view",
-        });
-        range.deleteContents();
-        range.insertNode($anchor[0]);
-        saveFile();
-      });
-    }
-  }
-};
+if (!$(input).length) {
+  console.error("input not found");
+}
 
-async function uploadImage() {
-  return new Promise((resolve, reject) => {
-    $input.attr({
-      type: "file",
+$(document).on("input", imageInput, async (e) => {
+  const fileID = $(fileSelector).val();
+  if (!fileID) {
+    console.error("Invalid file ID");
+    return;
+  }
+  const file = e.target.files[0];
+  if (!file) {
+    console.error("No file selected");
+    return;
+  }
+  const formData = new FormData();
+  formData.append("image", file);
+  const response = await fetch(`/upload/${fileID}`, {
+    method: "POST",
+    body: formData,
+  });
+  if (response.ok) {
+    const imageID = await response.text();
+    $("a.view.loading")
+      .attr("href", `/access/${fileID}/${imageID}`)
+      .removeClass("loading");
+    saveFile();
+  } else {
+    console.error(response.statusText);
+  }
+
+  $(this).val("");
+
+  const response2 = await fetch(`/access/${fileID}`);
+  if (response2.ok) {
+    const html = await response2.text();
+    console.log(html);
+    $(imagePanel).html(html);
+  }
+});
+
+const handleUploadImage = () => {
+  const selection = window.getSelection();
+  const selectedText = selection?.toString();
+  if (
+    selection &&
+    selection.rangeCount > 0 &&
+    selectedText &&
+    !/^\s+$/.test(selectedText)
+  ) {
+    const range = selection.getRangeAt(0);
+    const $anchor = $("<a>", {
+      href: `/res/image.svg`,
+      html: selectedText,
+      class: "view loading",
+    });
+
+    range.deleteContents();
+    range.insertNode($anchor[0]);
+  }
+
+  $(input)
+    .attr({
       name: "image",
       accept: "image/*",
-    });
-    $input.trigger("click").on("input", async () => {
-      $inputForm
-        .on("submit", async (e) => {
-          e.preventDefault();
+    })
+    .trigger("click");
+  $(contextMenu).trigger("complete");
+};
 
-          const formData = new FormData();
-          formData.append("image", $input[0].files[0]);
-          const fileID = $fileSelector.val();
-          try {
-            const response = await fetch(`/upload/${fileID}`, {
-              method: "POST",
-              body: formData,
-            });
-            if (!response.ok) {
-              throw new Error(response.statusText);
-            }
-            const text = await response.text();
-            console.log(text);
-            resolve(text); // Resolve the promise with the final result
-          } catch (err) {
-            console.log(err);
-            reject(err); // Reject the promise in case of an error
-          }
-        })
-        .trigger("submit");
-    });
-  });
-}
+$(uploadImage).on("click", async () => {
+  handleUploadImage();
+});
 
 export const uploadImageButton = () => {
   return $("<button>", {
-    id: "upload-image",
     text: "Upload Image",
-    class: "dropdown-item",
+    class: "dropdown-item upload-image",
   }).on("click", handleUploadImage);
 };