+++ frontend/public/js/readerHover.js
@@ -7,9 +7,9 @@ import {
   setMode,
   showMenuAtPos,
 } from "./contextMenu.js";
-import { $editor } from "./components.js";
+import { editor } from "./components.js";
 
-if (!$editor.length) {
+if (!$(editor).length) {
   console.error("Editor not found");
 }
 
@@ -22,7 +22,7 @@ const imageViewer = (src, alt) =>
     width: "50vw",
   });
 
-$editor.on("mouseover", "a.view", async (e) => {
+$(editor).on("mouseover", "a.view", async (e) => {
   if (getMode() === MODE.VIEW) {
     const anchor = e.target;
     const response = await fetch(anchor.href);