+++ frontend/public/js/titleSave.js
@@ -1,11 +1,11 @@
-import { $titles, $fileSelector } from "./components.js";
+import { title, fileSelector, renameFile, contextMenu } from "./components.js";
 
-if (!$titles.length) {
+if (!$(title).length) {
   console.error("No title found");
 }
 
 export async function saveTitle(filename) {
-  const fileID = $fileSelector.val();
+  const fileID = $(fileSelector).val();
   if (!fileID) {
     console.error("Invalid file ID");
     return;
@@ -23,17 +23,18 @@ export async function saveTitle(filename) {
 }
 
 $(document)
-  .on("click", ".title", (e) => {
+  .on("click", title, (e) => {
     e.preventDefault();
     const $title = $(e.target);
-    if ($title.text() === "New File") {
-      window.getSelection().selectAllChildren(e.target);
+    const selection = window.getSelection();
+    if ($title.text() === "New File" && selection) {
+      selection.selectAllChildren(e.target);
     }
     const target = e.target.closest(".id-selectable");
     if (!target) return;
-    $fileSelector.val(target.id);
+    $(fileSelector).val(target.id);
   })
-  .on("keypress", ".title", (e) => {
+  .on("keypress", title, (e) => {
     if (e.which === 13 || ((e.ctrlKey || e.metaKey) && e.key === "s")) {
       e.preventDefault();
       const $title = $(e.target);
@@ -46,6 +47,7 @@ $(document)
     if (e.which === 32) {
       e.preventDefault();
       const selection = window.getSelection();
+      if (!selection) return;
       const range = selection.getRangeAt(0); // Get current caret position
 
       // Insert a space character at the caret position
@@ -61,13 +63,23 @@ $(document)
       selection.addRange(range);
     }
   })
-  .on("paste", ".title", async (e) => {
+  .on("paste", title, async (e) => {
     e.preventDefault();
     const $title = $(e.target);
     const rawText = await navigator.clipboard.readText();
     const newFileName = rawText.trim().replace(/(\r\n|\n|\r)/gm, " ");
     $title.text(newFileName);
   })
-  .on("blur", ".title", (e) => {
+  .on("blur", title, (e) => {
     saveTitle($(e.target).text());
   });
+
+$(renameFile).on("click", () => {
+  const fileID = $(fileSelector).val();
+  if (!fileID) {
+    console.error("Invalid file ID");
+    return;
+  }
+  $(contextMenu).trigger("complete");
+  $(`#${fileID}`).find(title).trigger("focus");
+});