+++ frontend/public/js/trashPanel.js
@@ -1,15 +1,15 @@
-import { $contextMenu, $trashPanel, $fileSelector } from "./components.js";
+import { contextMenu, trashPanel, fileSelector } from "./components.js";
 import { appendItem, clearMenu, showMenuAtPos } from "./contextMenu.js";
 
-if (!$contextMenu.length) {
+if (!$(contextMenu).length) {
   console.error("Context menu not found");
 }
 
-if (!$trashPanel.length) {
+if (!$(trashPanel).length) {
   console.error("Trash panel not found");
 }
 
-if (!$fileSelector.length) {
+if (!$(fileSelector).length) {
   console.error("File selector not found");
 }
 
@@ -18,8 +18,8 @@ const handleFileSelect = (e) => {
   const target = e.target.closest(".id-selectable");
   if (!target) return;
   const fileID = target.id;
-  $fileSelector.val(fileID);
+  $(fileSelector).val(fileID);
   showMenuAtPos({ top: e.clientY, left: e.clientX });
 };
 
-$trashPanel.on("contextmenu", handleFileSelect).on("click", handleFileSelect);
+$(trashPanel).on("contextmenu", handleFileSelect).on("click", handleFileSelect);